fix docker-compose node_module caching
Previously node_modules were being generated in the image
and then the local directory was mounted over top. This resulted
in the local node_modules being used instead of those in the docker
image. Now we specify another override so that the container uses
the node_modules generated in the image.