Replace daphne with hypercorn/uvloop
Hypercorn is faster and doesn't use twisted underneath.

Unfortunately channels has a hard dependency on daphne for the dev
server so those deps can't be removed yet.

Installing uvloop causes bugs with channels (daphne, twisted) as twisted
tries to initialize an event loop when channels (daphne) is imported. To
remediate this, the channels app is only installed when in DEBUG mode.