Declare default Django auto field
A warning is emitted for models that do not have a primary key.

Django by default creates a primary key with an auto field but in 3.2
introduced a warning for when this happens implicitly without a
specified auto field class.

https://docs.djangoproject.com/en/3.2/releases/3.2/#customizing-type-of-auto-created-primary-keys

Let's just specify the default class used already by Django.