Remove docker swarm, use plain ol' docker-compose
Docker swarm is a pain to work with given the use case we have. Having
to build, publish and pull images is very resource intensive. Deploys
took on the order of 10 minutes, which for a basic app is quite silly.

The docker-compose overriding/inheriting was hard to reason about and
very easy to get wrong. Also, some docker-compose features were not
supported by docker stack/swarm which was very annoying.

The solution is to move to manual deploys with docker-compose and move
all configuration to the environment/.env files.

Now it's as easy as `docker-compose up -d` for local dev and
`docker-compose --env-file .env.prod up -d` for prod.

In the future it should be easy enough to set up webhooks to trigger a
git pull and docker-compose up to redeploy.